Move reactNativeTv to dependencies folder, fix test mocks
